Lift the polish orchestration into DictusCore, with one service per process (refs #361)
One PolishPipeline, two call sites (decision 4). PolishCoordinator's
orchestration -- the toggle, the language branch, the duration and gibberish
gates, cancellation, metrics -- becomes PolishService in DictusCore, and each
process owns an instance. That is not only code sharing: the #315 availability
gate is per-process by nature, because the rate limit it tracks is, and a gate
living in the app could say nothing true about the keyboard's budget
(decision 11).

The debug ring's file moves with it, as PolishEventStore. Both processes append;
only DictusApp prunes, rewrites, clears and exports, so every destructive
operation stays single-writer (decision 8). The keyboard's whole interaction is
append, which deliberately does not bootstrap -- a read of the seven-day window
would pull the entire file into a process under a ~50 MB ceiling. Appends go
through NSFileCoordinator for the reason PersistentLog does.

Every event now carries a writer marker, <KBD> or <APP>, spelled exactly as the
persistent log spells it so one grep answers the question in either artefact.
Without it the first export taken after this move is unreadable for the same
reason #315 needed appState: a clean export could equally mean the keyboard is
not rate-limited, or that little dictating happened. The export also gains a
per-writer failure breakdown, which is what makes decision 13's acceptance
criterion a thing a reader can check rather than infer.

Serialisation is explicit and says so: a superseded call logs how long the one
it displaced had been running (decisions 10 and 15).

PolishAvailabilityChannel changes owner. The notice describes the process that
calls the engine, and that is the keyboard now; leaving the app as writer would
put a permanent false "polish unavailable" line on a keyboard polishing
perfectly well, for as long as the app process lived.
